Add Raw Material Receiving (warehouse) feature: pages, models, API and routing
### Motivation

- Introduce a warehouse "采购入库" workflow to support blank instock creation, OCR ticket processing, QR binding and listing/canceling of raw material receiving orders.

### Description

- Add new warehouse models under `Models/Warehouse` (`RawMaterialReceivingDto`, `RawMaterialOcrDto`, `BlankInstockDto`, `WarehouseInfoDto`, `QrCodeInfoDto`, `AttachmentDto`, `DictGroupDto`) with display helper properties.
- Implement `IWarehouseApi` and `WarehouseApi` in `Services/Warehouse` to call new endpoints (`addBlankInstock`, `listInStock`, `queryWarehouseInfo`, `ocrIncoming`, `queryQrCodeInfo`, `getDictList`, `cancelBlankInstock`) and helpers for upload and response parsing.
- Add UI and behavior for the feature: `RawMaterialReceivingListPage` (list, refresh, cancel) and `AddRawMaterialReceivingPage` (capture/select photo, upload, OCR confirm dialog, scan bind flow, item summary) plus XAML and code-behind.
- Wire up routing and DI: register routes `RouteRawMaterialReceiving` / `RouteAddRawMaterialReceiving`, register pages in `MauiProgram.cs`, and expose module entry from `RoleHomePage` via `AppShell.RouteRawMaterialReceiving`.
- Update `Resources/Raw/appconfig.json` with new `rawMaterialReceiving` and `attachment` API endpoints and normalization in `WarehouseApi` via `IConfigLoader`.

### Testing

- No automated unit tests were added for this change; a local build was performed with `dotnet build` and the solution compiled successfully.

+++ Models/WorkOrders/MaterialQrCodeInfoDto.cs
@@ -12,7 +12,7 @@ public sealed class MaterialQrCodeInfoDto
     public string? qrCode { get; set; }
     public string? qrStatus { get; set; }
     public string? qrStatusName { get; set; }
-    public int? qrTimes { get; set; }
+    public decimal? qrTimes { get; set; }
     public string? sourceBizType { get; set; }
     public string? sourceBizTypeName { get; set; }
     public string? spec { get; set; }
